+CRUD +LOGIN
-Search

+++ app/src/main/java/com/example/dapenduk/ListPendudukActivity.java
@@ -1,16 +1,6 @@
 package com.example.dapenduk;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.lifecycle.ViewModelProviders;
-import androidx.recyclerview.widget.ItemTouchHelper;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
 import android.content.Intent;
-import android.net.sip.SipSession;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
@@ -18,6 +8,15 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.dapenduk.data.model.Penduduk;
 import com.example.dapenduk.ui.adapter.ListPendudukAdapter;
@@ -28,24 +27,23 @@ import java.util.List;
 
 public class ListPendudukActivity extends AppCompatActivity implements View.OnClickListener {
 
-    
 
     private PendudukViewModel pendudukViewModel;
-    private RecyclerView rvListPenduduk;
     private ListPendudukAdapter listPendudukAdapter;
-    private FloatingActionButton floatingActionButton,fabcount;
     private static final String TAG = "updatebos";
+    public static final int ADD_NOTE_REQUEST = 1;
+    public static final int EDIT_NOTE_REQUEST = 2;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_list_penduduk);
 
-        rvListPenduduk = findViewById(R.id.rv_penduduk);
-        floatingActionButton = findViewById(R.id.fab_add_penduduk);
-        fabcount = findViewById(R.id.fab_count_list);
+        RecyclerView rvListPenduduk = findViewById(R.id.rv_penduduk);
+        FloatingActionButton floatingActionButton = findViewById(R.id.fab_add_penduduk);
+        FloatingActionButton fabcount = findViewById(R.id.fab_count_list);
 
-        pendudukViewModel =  new ViewModelProvider(this).get(PendudukViewModel.class);
+        pendudukViewModel = new ViewModelProvider(this).get(PendudukViewModel.class);
         listPendudukAdapter = new ListPendudukAdapter();
 
         rvListPenduduk.setLayoutManager(new LinearLayoutManager(this));
@@ -58,14 +56,14 @@ public class ListPendudukActivity extends AppCompatActivity implements View.OnCl
             public void onChanged(List<Penduduk> penduduks) {
                 listPendudukAdapter.setPendudukList(penduduks);
                 Log.d(TAG, "onChanged: ");
-                Toast.makeText(ListPendudukActivity.this, "Onchange", Toast.LENGTH_SHORT).show();
+//                Toast.makeText(ListPendudukActivity.this, "Onchange", Toast.LENGTH_SHORT).show();
             }
         });
 
         floatingActionButton.setOnClickListener(this);
         fabcount.setOnClickListener(this);
 
-        new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0,ItemTouchHelper.LEFT) {
+        new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
             @Override
             public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
                 return false;
@@ -73,11 +71,26 @@ public class ListPendudukActivity extends AppCompatActivity implements View.OnCl
 
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                pendudukViewModel.delete(listPendudukAdapter.getpendudukPos(direction));
+                pendudukViewModel.delete(listPendudukAdapter.getpendudukPos(viewHolder.getAdapterPosition()));
                 Toast.makeText(ListPendudukActivity.this, "Deleted", Toast.LENGTH_SHORT).show();
                 pendudukViewModel.updateData();
             }
         }).attachToRecyclerView(rvListPenduduk);
+
+        listPendudukAdapter.setOnItemClickListener(new ListPendudukAdapter.OnItemClickListener() {
+            @Override
+            public void onItemClick(Penduduk penduduk) {
+                Intent intent = new Intent(ListPendudukActivity.this, AddEditPendudukActivity.class);
+                intent.putExtra(AddEditPendudukActivity.EXTRA_NAME, penduduk.getName());
+                intent.putExtra(AddEditPendudukActivity.EXTRA_ID, penduduk.getId());
+                intent.putExtra(AddEditPendudukActivity.EXTRA_ADDRESS, penduduk.getAddress());
+                intent.putExtra(AddEditPendudukActivity.EXTRA_BORN_AT, penduduk.getBornAt());
+                intent.putExtra(AddEditPendudukActivity.EXTRA_IS_MALE, penduduk.getIsMale());
+                intent.putExtra(AddEditPendudukActivity.EXTRA_PROFESSION, penduduk.getProfession());
+                startActivityForResult(intent, EDIT_NOTE_REQUEST);
+            }
+        }
+        );
 
     }
 
@@ -87,33 +100,68 @@ public class ListPendudukActivity extends AppCompatActivity implements View.OnCl
         pendudukViewModel.updateData();
     }
 
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if (requestCode == ADD_NOTE_REQUEST && resultCode == RESULT_OK && data != null) {
+            String name = data.getStringExtra(AddEditPendudukActivity.EXTRA_NAME);
+            String address = data.getStringExtra(AddEditPendudukActivity.EXTRA_ADDRESS);
+            String bornAt = data.getStringExtra(AddEditPendudukActivity.EXTRA_BORN_AT);
+            String profession = data.getStringExtra(AddEditPendudukActivity.EXTRA_PROFESSION);
+            boolean isMale = data.getBooleanExtra(AddEditPendudukActivity.EXTRA_IS_MALE, false);
+            Penduduk penduduk = new Penduduk();
+            penduduk.setName(name);
+            penduduk.setAddress(address);
+            penduduk.setBornAt(bornAt);
+            penduduk.setProfession(profession);
+            penduduk.setIsMale(isMale);
+
+            pendudukViewModel.insert(penduduk);
+            Toast.makeText(this, "Data Penduduk Saved", Toast.LENGTH_SHORT).show();
+        } else if (requestCode == EDIT_NOTE_REQUEST && resultCode == RESULT_OK && data!= null) {
+            long id = data.getLongExtra(AddEditPendudukActivity.EXTRA_ID, -1);
+
+            if (id == -1) {
+                Toast.makeText(this, "Pendduk can't be updated", Toast.LENGTH_SHORT).show();
+                return;
+            }
+
+            String name = data.getStringExtra(AddEditPendudukActivity.EXTRA_NAME);
+            String address = data.getStringExtra(AddEditPendudukActivity.EXTRA_ADDRESS);
+            String bornAt = data.getStringExtra(AddEditPendudukActivity.EXTRA_BORN_AT);
+            String profession = data.getStringExtra(AddEditPendudukActivity.EXTRA_PROFESSION);
+            boolean isMale = data.getBooleanExtra(AddEditPendudukActivity.EXTRA_IS_MALE, false);
+            Penduduk penduduk = new Penduduk(id, name, isMale, address, bornAt, profession);
+            pendudukViewModel.update(penduduk);
+        } else {
+            Toast.makeText(this, "Penduduk Not Saved", Toast.LENGTH_SHORT).show();
+        }
+    }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         MenuInflater menuInflater = getMenuInflater();
-        menuInflater.inflate(R.menu.list_menu,menu);
+        menuInflater.inflate(R.menu.list_menu, menu);
         return true;
     }
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()){
-            case R.id.delete_all:
-                pendudukViewModel.deleteAll();
-                Toast.makeText(this, "All Note Deleted", Toast.LENGTH_SHORT).show();
-                pendudukViewModel.updateData();
-        return true;
-        default:
-            return super.onOptionsItemSelected(item);
-
+        if (item.getItemId() == R.id.delete_all) {
+            pendudukViewModel.deleteAll();
+            Toast.makeText(this, "All Note Deleted", Toast.LENGTH_SHORT).show();
+            pendudukViewModel.updateData();
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
     public void onClick(View v) {
-        switch (v.getId()){
+        switch (v.getId()) {
             case R.id.fab_add_penduduk:
-                startActivity(new Intent(this, AddPendudukActivity.class));
+                startActivityForResult(new Intent(this, AddEditPendudukActivity.class), ADD_NOTE_REQUEST);
                 break;
             case R.id.fab_count_list:
                 pendudukViewModel.count();
